requirements.txt
Django==5.1
wagtail==6.2
psycopg2-binary
Pillow
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-modelcluster==6.3
django-taggit==5.0.1
djangorestframework
